moved the restaurant delete to the restaurant controller since its not returning any json objects
